test suite: normalize test case output when empty
It's more useful to ignore the differences between a completely empty output and
a single newline character.  We will never check the distinction between those
two outputs in the test suite, and being able to accept empty output is useful.

This solution is convenient, since the expected output is also generated by the
shell in a way that adds a final newline to any text.

* tests/utility.in(jitter_run_case_checking_output): Normalize test case output,
turning a completely empty output into a single newline character.