Makefile.am: fix mans dependency
Man pages, generated with help2man, do in practice depend on programs, even if
such a dependency is not explicitly expressed in the makefile as recommended
by the help2man documentation, for good reason.

The convenience "mans" target I added for the user built "all" before generating
the man pages.  This broke easily with parallel make, since man pages are
included in all.

* Makefile.am (mans): Build $(bin_PROGRAMS) $(bin_SCRIPTS) instead of all
as a first step before generating the man pages.