don't rewrite across caller instructions
A caller instruction is implicitly followed by a label describing the return
address for the callee.  Prevent rewriting from involving both a caller
instruction and what follows.

The code takes into account the possibility that a caller is added but then
replaced with a non-caller post-rewriting.  The added code is reachable at the
end of every rewritten instruction, so the check needs to be performed in just
one place.

* jitter/jitter-rewrite.c (jitter_rewrite): Here.