allow VM programs to end with a label
Define the concept of an unspecialized program "epilog", consisting of
unspecialized instructions to be added automatically just before specialization.
Right now the only added instruction is exitvm.  This plays the role of the old
(specialized) !EXITVM which was automatically added, and no longer is, and will
let me easily add more unspecialized instructions later as well, if needed.

The advantage of the new solution is that we can have a label at the very end of
an unspecialized program, and at the same time to keep every label associated to
exactly one unspecialized instruction.

!EXITVM still exists, but is now redundant.

* jitter/jitter-specialize.c (jitter_add_program_epilog): New static function,
which adds exitvm, a new implicit unspecialized instruction.
(jitter_specialize_program): Call jitter_add_program_epilog at the beginning
of the function.  Don't add !EXITVM.

* jitter/program.y: Uncomment grammar rules recognizing a program ending with a
label.  Remove the old version which prevented that.  The new set of recognized
programs is a strict superset of the old one.

* jitter/jitter-vm.h (struct jitter_vm): New field exitvm_meta_instruction.

* templates/vm1.c (vmprefix_initialize): Initialize
the_vmprefix_vm.exitvm_meta_instruction.