implement C hash table iteration
Provide a crude form of hash table iteration.  This is convenient for
JitterLisp.  Right now there is only a sort of fold function, calling a user
function on a key, a value, and a user pointer, walking thru the bindings in an
unspecified order.

* jitter/jitter-hash.h (jitter_hash_for_all_bindings_function): New function
type.
(jitter_hash_for_all_bindings): New prototype.

* jitter/jitter-hash.c (jitter_hash_for_all_bindings): New function.