don't rely on argp_error knowing how to print long long numbers
Don't use JITTER_PRIill to print an object which may be a long long using
a Gnulib function (argp_error) which might not support it.  I can do this
by losing no functionality, as the template in question was an error message
reproducing input from the user: I can just print the string.

* templates/vm-main.c (parse_opt): Print a string rather than a jitter_int.
Update copyright year.