fix the build system for non-standard prefix Jitter installations
Remove the configure-time option --with-jitter-config.

Change semantics for the configure-time option --with-jitter , using a prefix
instead of the path of a bin directory (suggested by José Marchesi).  Use
the same prefix for jitter-config as well.

jitter-config: add a -I option to the output of --cppflags .
jitter-config: change semantics for the option --include-path , now printing
the path of the header directory containing jitter/ , rather than of the
directory named jitter/ .  This new behavior is more useful.

Update the structured build system used as an example, for the user's sake.

* Makefile.am (example-vms/structured/build-aux/jitter.m4): Make a copy in
the example subdirectory, just to present something reasonable to the user.
Distribute the copy as well.

* autoconf/jitter.m4 (AC_JITTER_CONFIG): Remove --with-jitter-config option.
Support --with-jitter here, with the new semantics.  Comment changes.
(AC_JITTER_C_GENERATOR): AC_PATH_PROG: add $PATH as fallback.
Update copyright years.

* example-vms/structured/configure.ac: Use AC_CONFIG_MACRO_DIR to make
the copy of jitter.m4 visible from configure.  This is the recommended
solution for users.

* scripts/jitter-config.in.m4sh (includedir): Redefine not to include the
"jitter" subdirectory.
(help): Update help to describe the new semantics for --include-path.
(Main loop): prepend "-I $includedir" to the output of --cppflags.
Update copyright years.