add a third way for the user to add a VM instruction, by its unspecialized id
The current two ways for the user of adding a VM instruction to a program are
VMPREFIX_APPEND_INSTRUCTION, which requires the instruction name to be known
at compile time, and vmprefix_append_instruction_name, which is more flexible
but less efficient.

Add VMPREFIX_APPEND_INSTRUCTION_ID as a third way, taking a C expression
evaluating to an unspecialized instruction id (as an enum case) as an argument.

* jitter/jitter-program.h (jitter_append_instruction_id): New prototype.

* jitter/jitter-program.c (jitter_append_instruction_id): New function.

* templates/vm.h (VMPREFIX_APPEND_INSTRUCTION_ID): New macro.
(VMPREFIX_APPEND_INSTRUCTION): Update comment, comparing with the new macro as
well.

Original idea from a discussion with José Marchesi.