jitterlisp: optimize tagged plus and minus
On two's complement machines, which are the only machines supported by
JitterLisp (and GNU) we can have the guarantee that the sum or subtraction of
two tagged fixnums with a tag of 0 in low-order bits is itself automatically
tagged the right way.  Of course I am currently ignoring overflow.

No longer explicitly tag the result of a fixnum sum or subtraction.  Now a
fixnum + or - takes one machine instruction (without counting type checks), as
long as the machine instructions for summing and subtracting allow 16 as an
immediate.

Thanks to José Marchesi for confirming to me, with a convincing mathematical
argument, that this optimization is actually correct.

* example-vms/jitterlisp/jitterlisp-operations.h
(JITTERLISP_EXP_FF_F_PLUS_OR_MINUS) [fixnum tag zero]: Redefine.
Add comment about multiplication, which can probably benefit from a similar,
if slightly less dramatic, optimization.