actually implement the option slow_literals_only
Add a new boolean parameter to Jitter-generated instruction recognizers.  This
parameters tells whether fast literals are enabled or not.  The rest of the API
does not need to change: the first call to the recognizer receives a VM program
pointer as an argument, and the program contains the option slow_literals_only.

Specializer functions don't need any change: all the relevant decisions are
taken by recognizer functions.

* jitterc/jitterc-generate.c (jitterc_emit_specializer_recognizer_prototypes):
Add boolean parameter in generated prototypes for
vmprefix_recognize_specialized_instruction_INSTRUCTION.
(jitterc_emit_specializer_recognizers): Add boolean argument in generated
recognizer functions.  Use it in generated code in && with value comparison
in the non-residual literal case, so as to make the condition never fire when
fast literals are disabled.  At the end of the generated recognizer function,
when calling the recognizer for the next VM instruction parameter, pass the
boolean parameter unchanged.
(jitterc_emit_specializer): Extract the boolean from the program option.  Pass
it to the first recognizer.