build system: no longer distribute jitter-generated C files
In an ordinary application it would make sense to distribute generated files as
well, as per the ordinary GNU convention; but here we are in the build system of
Jitter itself, which allows to rebuild those files.  And they are only examples,
so they are not crucial to build, nor indeed built by default.  If the user is
able to run the examples or the testsuite, then there is some emulator allowing
her to regenerate the files as well.

* Makefile.am (EXTRA_DIST): Do not add $(JITTER_GENERATED_FILES).  Explain why
in a long comment.