no-threading: avoid jitter_ip except at initialization, freeing a register
* jitterc/jitterc-generate.c (jitterc_emit_executor_main_function)
[NO_THREADING]: In emitted code, remove commented-out CPP conditional around
the declaration of jitter_ip.  Later, in emitting code to branch to the first
VM instruction, emit a comment and emit a macro definition making jitter_ip
an alias for vmprefix_array_base_register_variable, except when jumping to
the first VM instruction.  This kills the variable soon, and in my tests frees
up one hardware register which would be completely wasted on no-threading;
occasionally this even avoids spurious register moves.

* jitter/jitter-executor.h (JITTER_PRETEND_TO_POSSIBLY_JUMP_TO_): Rename
assembly operand from jitter_ip to to the_jitter_ip, so that the new CPP
expansion does not generate something in C which does not match the
assembly template.