configure: fix bug; check for GNU C attributes
When checking for the presence of __builtin_add_overflow_p and other builtins
I have to actually link; compiling alone is not enough to check whether the
compiler knows about the builtins.

Check for the presence of GNU C attributes, and (at this time) one specific
attribute.

* configure.ac [JITTER_HAVE_GCC_OVERFLOW_CHECKING]: Use AC_LINK_IFELSE instead
of AC_COMPILE_IFELSE.
[JITTER_HAVE_GNU_C_ATTRIBUTE, JITTER_HAVE_ATTRIBUTE_RETURNS_NONNULL]: New
checks.  Not used yet.