disable test suite by default in sub-package mode
Rename the configure option --enable-test-suite to --enable-jitter-test-suite,
in order to prevent collisions with configuration option for the super-package
in sub-package mode.

By default disable Jitter's test suite in sub-package mode.  In sub-package
mode, when enabled, Jitter's test suite would run as a side effect of running
the test suite of the super-package, which would be time consuming and
distracting.

Now, in sub-package mode with Jitter's test suite disabled, attempting to run
the test suite anyway by means of make check or make recheck will not fail, but
will not actually run the tests either, or print distracting output.

* configure.ac (Test suite support, default): Change to no in sub-package mode.
Update comments and user messages.
(test-suite): Rename configure option --enable-test-suite to
--enable-jitter-test-suite.  Update help string.
(TEST_SUITE_LOG): Define as a substitution in order to override the Automake
default.  Explain why in a comment.

* Makefile.am (Test suite) [! JITTER_ENABLE_TEST_SUITE]: Add comment, explaining
behavior in sub-package mode with the test suite disabled.  No change needed in
the actual code.