do not automatically trap or break on division or remainder by 0 on MIPS
GCC defaults to generate a conditional trap or break instruction after divisions
on MIPS, which is wasteful.  In C division and remainder by zero have undefined
semantics anyway, and safe languages compiled to VM code will need to add their
own check in addition to this.

Do not generate a trap-on-zero or break-on-zero instruction at division time on
MIPS, with any dispatch.

* configure.ac (Flags) [MIPS]: Check for availability of the compiler option
-mno-check-zero-division, and use it.
(Flags): Unrelated comment indentation fix.