fix make tags, working around Automake bug 38139
'make tags' failed in Jitter because of this Automake bug:
  https://debbugs.gnu.org/cgi/bugreport.cgi?bug=38139
  https://lists.gnu.org/archive/html/automake/2019-10/msg00001.html

I reported the bug along with a trivial one-character fix and Karl Berry
provisionally acknowledged the problem.  The Automake maintainers have not
responded yet, but that is not terribly important: older Automake installations
will remain in use for some time, so it is better to work around the issue and
ensure compatibility with all versions.  Do that.

* Makefile.am (dist_lisp_LISP): Replace with dist_lisp_DATA, to work around
an Automake bug.  Add comments.