new stack operation: slide
Add a new stack operation, slide.  This is a generalization of nip allowing to
remove a given number of elements at a given depth from a TOS or non-TOS stack.
The operation is fast when the arguments are known at compile time, but can be
dangerous if used carelessly: replacing the top is forbidden.

Add VM instructions for slide in the Uninspired example.

* jitter/jitter-stack.h (JITTER_STACK_TOS_SLIDE, JITTER_STACK_NTOS_SLIDE): New
macros.

* jitterc/jitterc-generate.c (jitterc_emit_stack_operation_definitions): Emit
friendly user definitions for the slide operation.

* example-vms/uninspired/uninspired.jitter (slide, slideforth): New VM
instructions.

* emacs/jitter-mode.el (jitter-font-lock-keywords): Add font lock support for
JITTER_SLIDE_.