unified routines: fix bug in routine printing
I forgot to automatically make an executable routine whenever a unified routine
is printed.  This caused a surprising bug in Structured, where routines worked
correctly and disassembled to correct native code but printed with misplaced
labels.  Labels are resolved at the time an executable routine is made.

Poke should not be affected, as I took care of explicitly generating executable
routines as soon as possible when I ported it to the unified routine API, in
order to prevent reentrancy problems later when a routine is executed for the
first time.

JitterLisp is not affected because it does not use the unified API.  Likewise,
machine-generated frontends for VMs do not use the unified API.

Do not make jitter_routine_print just a macro alias for
jitter_mutable_routine_print: turn it into a real C function doing the missing
part of the job.  Fix macro wrappers.

No visible changes in the user API.

* jitter/jitter-routine.h (jitter_routine_print): New prototype.

* jitter/jitter-routine.c (jitter_routine_print): New function.

* templates/vm.h (vmprefix_routine_print): Redefine as a macro alias for
jitter_routine_print instead of jitter_mutable_routine_print.  Now that
vmprefix_routine_print fits the same pattern as vmprefix_disassemble_routine
and vmprefix_disassemble_routine_to merge the comments for all three.