portability fix: do not insert \n with sed
Some weak versions of sed do not insert a newline character in the place of \n
when \n occurs in the replacement part.  This surprising behavior broke Jitter's
test suite machinery on at least one platform, where even correct output never
compared as equal to the expected result.

Fix by not relying on sed expanding \n as a newline character.

* tests/utility.in (jitter_run_case_checking_output): Use sed and tr rather than
just sed.  Forbid the '$' character (now used in the intermediate step) in
expected output.  Add comment.

Thanks to Niobos for reporting and running my test code to help me to locate the
issue.