human prefixes: fixes and improvements
Change the internal algorithm to make sure that every number is printed with the
smallest prefix which is strictly larger than it, or the smallest prefix if the
number is smaller than the smallest prefix, or the largest prefix if the number
is larger than the largest prefix.  Simplify the algorithm.  Minor cleanups.

* jitter/jitter-human-prefix.c (jitter_human_readable_with): Completely change
the algorithm.  Whitespace.
(jitter_human_readable): Rename formal from scale to prefix.  Whitespace.

* jitter/jitter-human-prefix.h (JITTER_HUMAN_READABLE_): Comment fix.