human readable prefix fixes
Fix a bug due to which 0 was given a yocto- prefix.

Remove binary fractional prefixes altogether.  They are not IEC-standard and,
even if fractional prefixes are useful for averages, they are not in wide use.

One byte every 1000 seconds could still maybe be called a millibyte and given
the prefixed unit "mB", but it is hard to see a case where calling one byte
every 1024 seconds a mibibyte ("miB") would be of any use.  Binary powers are
useful for expressing measures in relation to quantities which are nonnegative
powers of two, not really for fractions.

No API changes.

* jitter/jitter-human-prefix.c (jitter_human_descriptor_binary): Remove
descriptors for fractional binary prefixes.
(jitter_human_readable): Fix the algorithm by changing in which case the
function returns.