aligned block: fix the aligned_alloc version following Bruno Haible's bug report
Bruno let me know that aligned_alloc requires the allocated size to be a
multiple of the required alignment.  This restriction is explicitly documented
in the GNU libc manual, but had escaped me.

Round the size up: the aligned block implementation based on aligned_alloc will
be a little more wasteful (in theory: in practice the alignment is the same as
the size, the way I use aligned blocks from the garbage collector), but now
correct.

When deciding on which implementation to use, give priority to posix_memalign
which lacks this restriction.  mmap is still preferable to either.

* jitter/jitter-aligned-block.c (jitter_aligned_block_make)
[JITTER_ALIGNED_BLOCK_USE_ALIGNED_ALLOC]: Round size_in_bytes up.  Add comment.

* jitter/jitter-aligned-block.h (JITTER_ALIGNED_BLOCK_USE_ALIGNED_ALLOC)
[JITTER_HAVE_ALIGNED_ALLOC]: Reorder CPP conditional cases: in light of the
restriction, prefer posix_memalign to aligned_alloc.
Update the comment explaining the various implementations, mentioning the
problem.

Thanks to Bruno Haible for his bug report.