disassemble: work around a qemu problem, mostly for my own testing
* jitter/jitter-disassemble.c (jitter_disassemble_range_objdump): Return success
when pclose fails with error ENOSYS.  This is benign, and the underlying
operation in fact succeeds.
I began observing this misbehaviour on riscv32 (not riscv64) when running under
qemu-user since updating qemu from git, yesterday.  No other architecture I test
is affected.  I may want to remove this workaround in the future.