make JITTER_PRETEND_TO_POSSIBLY_JUMP_ANYWHERE usable with switch dispatching
I am about to use it, in the next change set.  The fact that the expansion
requires GNU C will not be a problem.

* jitter/jitter-executor.h [JITTER_DISPATCH_SWITCH] (jitter_dispatch_label):
Define as a macro, for compatibility with the other dispatches, to be used...
(JITTER_PRETEND_TO_POSSIBLY_JUMP_ANYWHERE): ...Here.  Rename the asm operand
from jitter_ip to _jitter_ip, to prevent macroexpansion.