fix jitter_dispatch_label bug
Make sure that a label jitter_dispatch_label is defined with any dispatch, as
required by profiling instrumentation.

This fixes a bug reported by José Marchesi.  Thanks José.

* jitter/jitter-executor.h [JITTER_DISPATCH_SWITCH] (jitter_dispatch_label):
Remove the macro.  Change it into a real label...

* jitterc/jitterc-generate.c (jitterc_emit_executor_main_function): ...Here.
Emit the jitter_dispatch_label label unconditionally, even if it is in fact
always unreachable.  For switch dispatch make it branch (goto) to the real
dispatch label.  For the other dispatches it branch (goto *) to ip.
Of course there is no real threaded ip with no-threading, but as
usual I fool the compiler into using the base as if it were ip,
in order to have something with an unpredictable value, and without using any
other register.