ios: Fix the behaviour of the flags in file IOD
The fopen modes (r, w, a, r+, w+, a+) do not handle all possibilities
that we want to support via IOS_F_READ, IOS_F_WRITE and IOS_F_CREATE.
To solve this issue, this patch replaces calls to fopen with calls to
open and fdopen. Although fdopen uses the same set of modes, it skips
the implications of those modes regarding file creation and truncation.

2021-03-15  Egeyar Bagcioglu  <egeyar@gmail.com>

        * libpoke/ios-dev-file.c (ios_dev_file_convert_flags): New function.
        (ios_dev_file_open): Open files with the right flags via "open".