poke,ios: dump should not print anything if the IOS is not readable
This patch makes dump to check that the given IOS exists and is
readable before trying to emit any output (like ruler and/or ASCII).

Note the XXX note:

  /* First of all, we require a readable IO space.  */
  if (! (ioflags (ios) & IOS_F_READ))
    /* XXX change to E_perm when it becomes available.  */
    raise E_no_ios;

This will be changed as soon as we have support for E_perm, which has
been sent to poke-devel for review.

Note also that this patch makes the memory IOD to ignore user-provided
flags and set them to read/write.

2021-03-27  Jose E. Marchesi  <jemarch@gnu.org>

	* libpoke/ios-dev-mem.c (ios_dev_mem_open): Memory IOS are always
	read/write.
	* poke/pk-dump.pk (Arguments): Raise E_perm if the specified IOS
	is not readable.
	* testsuite/poke.cmd/dump-10.pk: New test.
	* testsuite/poke.cmd/dump-11.pk: Likewise.
	* testsuite/Makefile.am (EXTRA_DIST): Add new test.