pickles: btf: improve magic num endianness logic
The magic number should always read as 0xeb9f (like eBPF). When mapping
a BTF_Header, if 0x9feb is read instead then the current endianness is
simply the opposite that of the BTF information.

This fixes an issue where attempting to map a little-endian BTF_Header
while in big-endian mode, the endianness would not correctly be changed.

2022-01-04  David Faust  <david.faust@oracle.com>

	* pickles/btf.pk (BTF_Header): Adjust constraint on magic number
	to properly set endianness as needed.